moved project to Sources folder
