+++ Package.swift
@@ -27,5 +27,6 @@ let package = Package(
         .testTarget(
             name: "RyeTests",
             dependencies: ["Rye"]),
-    ]
+    ],
+    swiftLanguageVersions: [.v4_2, .v5]
 )
