--- Sources/Rye/Rye.swift
@@ -1,3 +1,0 @@
-struct Rye {
-    var text = "Hello, World!"
-}
