+++ Tests/RyeTests/RyeTests.swift
@@ -1,15 +1,34 @@
+//
+//  RyeTests.swift
+//  RyeTests
+//
+//  Created by Andrei Hogea on 12/06/2019.
+//  Copyright © 2019 Nodes. All rights reserved.
+//
+
 import XCTest
 @testable import Rye
 
-final class RyeTests: XCTestCase {
-    func testExample() {
-        // This is an example of a functional test case.
-        // Use XCTAssert and related functions to verify your tests produce the correct
-        // results.
-        XCTAssertEqual(Rye().text, "Hello, World!")
+class RyeTests: XCTestCase {
+
+    override func setUp() {
+        // Put setup code here. This method is called before the invocation of each test method in the class.
     }
 
-    static var allTests = [
-        ("testExample", testExample),
-    ]
+    override func tearDown() {
+        // Put teardown code here. This method is called after the invocation of each test method in the class.
+    }
+
+    func testExample() {
+        // This is an example of a functional test case.
+        // Use XCTAssert and related functions to verify your tests produce the correct results.
+    }
+
+    func testPerformanceExample() {
+        // This is an example of a performance test case.
+        self.measure {
+            // Put the code you want to measure the time of here.
+        }
+    }
+
 }
